refactor(study-tracking): move domain modules
