+++ server/routes/admin-reports.fastify.ts
@@ -37,7 +37,7 @@ import {
   createRuntimeTimer,
   type RuntimeTimer,
 } from "../lib/runtime-timing.ts";
-import { ensureStudyTrackingCaseForToken } from "../lib/token-study-tracking.ts";
+import { ensureStudyTrackingCaseForToken } from "../features/study-tracking/domain/index.ts";
 
 type AdminUserRecord = {
   id: number;
