+++ server/routes/particular-study-tracking.fastify.ts
@@ -23,7 +23,7 @@ import {
   parsePositiveInt,
   serializeStudyTrackingCase,
   serializeStudyTrackingNotification,
-} from "../lib/study-tracking.ts";
+} from "../features/study-tracking/domain/index.ts";
 import {
   buildRequestLogLine,
   sanitizeUrlForLogs,
