+++ test/unit/contracts/study-tracking/token-study-tracking.test.ts
@@ -2,8 +2,10 @@ import test from "node:test";
 import assert from "node:assert/strict";
 
 import type { StudyTrackingCase } from "../../../../drizzle/schema.ts";
-import { calculateEstimatedDeliveryAt } from "../../../../server/lib/study-tracking.ts";
-import { ensureStudyTrackingCaseForToken } from "../../../../server/lib/token-study-tracking.ts";
+import {
+  calculateEstimatedDeliveryAt,
+  ensureStudyTrackingCaseForToken,
+} from "../../../../server/features/study-tracking/domain/index.ts";
 
 function createStudyTrackingCaseFixture(
   input: Omit<StudyTrackingCase, "id" | "createdAt" | "updatedAt">,
