+++ test/unit/domain/study-tracking/study-tracking-clinic-schema.test.ts
@@ -3,7 +3,7 @@ import assert from "node:assert/strict";
 import {
   STUDY_TRACKING_STAGES,
   clinicCreateStudyTrackingSchema,
-} from "../../../../server/lib/study-tracking.ts";
+} from "../../../../server/features/study-tracking/domain/index.ts";
 
 test("STUDY_TRACKING_STAGES conserva el orden público esperado", () => {
   assert.deepEqual(STUDY_TRACKING_STAGES, [
