+++ test/unit/domain/study-tracking/study-tracking.test.ts
@@ -22,7 +22,7 @@ import {
   serializeStudyTrackingNotification,
   shouldCreateSpecialStainNotification,
   updateStudyTrackingSchema,
-} from "../../../../server/lib/study-tracking.ts";
+} from "../../../../server/features/study-tracking/domain/index.ts";
 
 test("adminCreateStudyTrackingSchema normaliza booleanos, textos y fechas", () => {
   const parsed = adminCreateStudyTrackingSchema.safeParse({
